Properties 분리
